+++ packages/config/package.json
@@ -45,6 +45,8 @@
   "devDependencies": {
     "@scalar/build-tooling": "workspace:*",
     "@types/node": "^20.14.10",
-    "rollup": "^4.16.4"
+    "rollup": "^4.16.4",
+    "vite": "^5.2.10",
+    "vitest": "^1.6.0"
   }
 }
