+++ packages/config/src/configuration.ts
@@ -1,3 +1,3 @@
-import ScalarConfig from '../schemas/config.json'
+import ScalarConfig from './schemas/config.json'
 
 export const ScalarConfigSchema = ScalarConfig
