+++ packages/config/src/schemas/config.json
