+++ packages/config/tsconfig.build.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "include": ["src"],
-  "exclude": ["dist", "test", "**/*.test.ts"],
+  "exclude": ["dist", "test", "vite.config.ts", "**/*.test.ts"],
   "compilerOptions": {
     "skipLibCheck": true,
     "declaration": true,
