+++ packages/config/vite.config.ts
@@ -1,0 +1,18 @@
+import { createViteBuildOptions, findEntryPoints } from '@scalar/build-tooling'
+import { URL, fileURLToPath } from 'node:url'
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  plugins: [],
+  resolve: {
+    alias: {
+      '@': fileURLToPath(new URL('./src', import.meta.url)),
+    },
+  },
+  server: {
+    port: 9000,
+  },
+  build: createViteBuildOptions({
+    entry: await findEntryPoints({ allowCss: false }),
+  }),
+})
